move is_cfg_view from delta.py to types.py
