+++ edb/pgsql/delta.py
@@ -114,11 +114,6 @@ if TYPE_CHECKING:
     from edb.schema import schema as s_schema
 
 
-# Modules where all the "types" in them are really just custom views
-# provided by metaschema.
-VIEW_MODULES = ('sys', 'cfg')
-
-
 def has_table(obj: Optional[so.InheritingObject], schema: s_schema.Schema):
     assert obj
 
@@ -150,31 +145,6 @@ def has_table(obj: Optional[so.InheritingObject], schema: s_schema.Schema):
             ptr_stor_info is not None
             and ptr_stor_info.table_type == 'link'
         )
-
-
-def is_cfg_view(
-    obj: so.Object,
-    schema: s_schema.Schema,
-) -> bool:
-    return (
-        isinstance(obj, (s_objtypes.ObjectType, s_pointers.Pointer))
-        and (
-            obj.get_name(schema).module in VIEW_MODULES
-            or bool(
-                (cfg_object := schema.get(
-                    'cfg::ConfigObject',
-                    type=s_objtypes.ObjectType, default=None
-                ))
-                and (
-                    nobj := (
-                        obj if isinstance(obj, s_objtypes.ObjectType)
-                        else obj.get_source(schema)
-                    )
-                )
-                and nobj.issubclass(schema, cfg_object)
-            )
-        )
-    )
 
 
 DEFAULT_INDEX_CODE = ' ((__col__) NULLS FIRST)'
@@ -303,7 +273,7 @@ class MetaCommand(sd.Command, metaclass=CommandMeta):
         assert isinstance(ctx.op, CompositeMetaCommand)
 
         src = ptr.get_source(schema)
-        if src and is_cfg_view(src, schema):
+        if src and types.is_cfg_view(src, schema):
             assert isinstance(src, s_sources.Source)
             self.pgops.add(
                 CompositeMetaCommand._refresh_fake_cfg_view_cmd(
@@ -2048,7 +2018,7 @@ class ConstraintCommand(MetaCommand):
         ):
             return False
 
-        if is_cfg_view(subject, schema):
+        if types.is_cfg_view(subject, schema):
             return False
 
         match subject:
@@ -3075,7 +3045,7 @@ class CompositeMetaCommand(MetaCommand):
 
     @staticmethod
     def _get_table_name(obj, schema) -> tuple[str, str]:
-        is_internal_view = is_cfg_view(obj, schema)
+        is_internal_view = types.is_cfg_view(obj, schema)
         aspect = 'dummy' if is_internal_view else None
         return common.get_backend_name(
             schema, obj, catenate=False, aspect=aspect)
@@ -3158,7 +3128,7 @@ class CompositeMetaCommand(MetaCommand):
         cols = []
 
         special_cols = ['tableoid', 'xmin', 'cmin', 'xmax', 'cmax', 'ctid']
-        if not is_cfg_view(obj, schema):
+        if not types.is_cfg_view(obj, schema):
             cols.extend([(col, col, True) for col in special_cols])
         else:
             cols.extend([('NULL', col, False) for col in special_cols])
@@ -3293,7 +3263,10 @@ class CompositeMetaCommand(MetaCommand):
             # figure that out, do *something* so that DDL isn't
             # excruciatingly slow because of the cost of explicit id
             # checks. See #5168.
-            and (not is_cfg_view(child, schema) or is_cfg_view(obj, schema))
+            and (
+                not types.is_cfg_view(child, schema)
+                or types.is_cfg_view(obj, schema)
+            )
         ]
 
         # Hackily force 'source' to appear in abstract links. We need
@@ -3331,7 +3304,7 @@ class CompositeMetaCommand(MetaCommand):
         context: sd.CommandContext,
         obj: CompositeObject,
     ) -> None:
-        if is_cfg_view(obj, schema):
+        if types.is_cfg_view(obj, schema):
             self._refresh_fake_cfg_view(obj, schema, context)
 
         bases = set(obj.get_bases(schema).objects(schema))
@@ -3410,7 +3383,7 @@ class CompositeMetaCommand(MetaCommand):
     ) -> None:
         assert has_table(obj, schema)
 
-        if is_cfg_view(obj, schema):
+        if types.is_cfg_view(obj, schema):
             self._refresh_fake_cfg_view(obj, schema, context)
 
         inhview = self.get_inhview(schema, obj, exclude_ptrs=exclude_ptrs)
@@ -3436,7 +3409,7 @@ class CompositeMetaCommand(MetaCommand):
     ) -> None:
         assert has_table(obj, schema)
 
-        if is_cfg_view(obj, schema):
+        if types.is_cfg_view(obj, schema):
             self._refresh_fake_cfg_view(obj, schema, context)
 
         inhview = self.get_inhview(
@@ -3888,8 +3861,8 @@ class ObjectTypeMetaCommand(AliasCapableMetaCommand, CompositeMetaCommand):
         # configs, since those need to be created after the standard
         # schema is in place.
         if not (
-            is_cfg_view(scls, eff_schema)
-            and scls.get_name(eff_schema).module not in VIEW_MODULES
+            types.is_cfg_view(scls, eff_schema)
+            and scls.get_name(eff_schema).module not in types.VIEW_MODULES
         ):
             return
 
@@ -5773,7 +5746,7 @@ class PropertyMetaCommand(PointerMetaCommand[s_props.Property]):
                 (default := prop.get_default(schema))
                 and not prop.is_pure_computable(schema)
                 and not fills_required
-                and not is_cfg_view(src.scls, schema)  # sigh
+                and not types.is_cfg_view(src.scls, schema)  # sigh
                 # link properties use SQL defaults and shouldn't need
                 # us to do it explicitly (which is good, since
                 # _alter_pointer_optionality doesn't currently work on
@@ -6147,7 +6120,7 @@ class UpdateEndpointDeleteActions(MetaCommand):
             x for obj in objs for x in obj.descendants(schema)}
         return {
             obj for obj in objs
-            if not obj.is_view(schema) and not is_cfg_view(obj, schema)
+            if not obj.is_view(schema) and not types.is_cfg_view(obj, schema)
         }
 
     def get_orphan_link_ancestors(self, link, schema):
@@ -6704,7 +6677,7 @@ class UpdateEndpointDeleteActions(MetaCommand):
 
             if (
                 not isinstance(source, s_objtypes.ObjectType)
-                or is_cfg_view(source, eff_schema)
+                or types.is_cfg_view(source, eff_schema)
             ):
                 continue
 
@@ -6753,7 +6726,7 @@ class UpdateEndpointDeleteActions(MetaCommand):
         delete_target_targets = set()
 
         for target in all_affected_targets:
-            if is_cfg_view(target, schema):
+            if types.is_cfg_view(target, schema):
                 continue
 
             deferred_links = []
@@ -6782,7 +6755,7 @@ class UpdateEndpointDeleteActions(MetaCommand):
                 source = link.get_source(schema)
                 if (
                     not source.is_material_object_type(schema)
-                    or is_cfg_view(source, schema)
+                    or types.is_cfg_view(source, schema)
                 ):
                     continue
